Split relocation tests into multiple files
This avoids requiring kernel support for all types of relocations when
testing a specific type.
